refactor: recoil 코드 분리
