+++ src/store/darkmode.ts
@@ -1,9 +1,4 @@
-import { atom, DefaultValue, selector } from 'recoil'
-
-export interface IToastState {
-  visible: boolean
-  detail: string
-}
+import { DefaultValue, atom, selector } from 'recoil'
 
 interface IDarkmodeState {
   darkmode: boolean
@@ -30,26 +25,6 @@ export const darkmodeSelector = selector<boolean>({
 
     if (!(value instanceof DefaultValue)) {
       set(darkmodeState, { ...get(darkmodeState), darkmode: value })
-    }
-  },
-})
-
-export const toastState = atom<IToastState>({
-  key: 'toastState',
-  default: {
-    visible: false,
-    detail: '',
-  },
-})
-
-export const toastSelector = selector<IToastState>({
-  key: 'modalSelector',
-  get: ({ get }) => get(toastState),
-  set: ({ set, reset }, value) => {
-    if (value instanceof DefaultValue) {
-      reset(toastState)
-    } else {
-      set(toastState, { ...value })
     }
   },
 })
